requirements.txt
streamlit>=1.28.1
langgraph>=0.0.32
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
pandas>=2.1.3
plotly>=5.17.0
openai>=1.3.5
python-dotenv>=1.0.0
numpy>=1.24.0
